requirements.txt
appdirs==1.4.4
distlib==0.3.1
Django==2.2.15
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
Markdown==3.2.2
mysqlclient==1.4.6
psycopg2==2.8.5
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
virtualenv==20.0.31
